fix generic/method mismatch in predict
Method used first argument hM instead of generic object. Most
references to hM in the code were changed with sed 's/hM\$/object$/g'
(and this may need checking).

There is still one doc discrepancy: rL was documented, but not used.
Probably a relic of an older version of the function.