Future R 4.0 has class c("matrix","array"): failure in R-devel check
R CMD check --as-cran fail now for R-devel (R 4.0-to-be) which
changes the class of matrix from "matrix" to c("matrix","array"),
and we have conditions which assume length-1 output from class().

This is the first function to fail in checks, but there are others
that fail when checked after this is fixed, and many functions need
checking although they do not fail.

Most (63 of 120) tests in the tests/ directory also fail.