copy current README.md to inst/
The purpose is to customize a README.md for the CRAN package. For
this we need to remove github instructions and dynammic badges
(travis test status, CRAN release version, downloads). Ideally this
could be something similar as man/Hmsc-package.Rd: this should be
seen as the first step towards more useful CRAN README (as CRAN
also exposes the README on its package front page).