verbose must be an integer or action can vary with a platform
B.D.Ripley notified us that example(sampleMcmc) did not match the
saved tests/Examples/Hmsc-Ex.Rout.save. The reason for inconsistency
was that the default was verbose = samples*thin/100 which for
samples=20 was 0.2. In my main platform (1:20) %% 0.2 was always 0
and every sampling was reported, but in most other platforms this
was never 0 and nothing was reported, and the test reported changed
output.