prettier diagnostic messages: do not use print()
print() appends every line with [1] and quotes the message. Now use
cat() for simple tracing info, message() for updater warnings, and
warning() when it looks appropriate.

Currently we have internal variable 'updaterWarningFlag' to turn
messages on or off, but that is not exposed to users. With message()
users can suppress messages if they wish.