rebuild vignette 5 after changes in its source and in Hmsc
Vignette 5 was added on Dec 17, 2019 and never updated later. After
there was a change in its source (Rmd) file (7b24b32) I decided to
re-build the pdf. For this the Rmd needed further editing so that
it can be built. There are also changes in Hmsc which will influence
the results. Most important change is db7cf523 that changes priors
for normal and lognormal poisson models that change mixing properties.
Minor fix in 7b24b32 may also influence the results, and probably
there are some other minor changes that can influence the pdf.