allow use of SpatialPoints as spatial coordinates
Currently implemented only for Full spatial model: NNGP & GPP
need to work with original spatial coordinatesö and they use
functions that are not directly available for SpatialPoints.
NNGP could be chnaged to handle distances, and both NNGP and GPP
could work with projected coordinates, but currently we do not
allow these.