find diagonal of SpatialPoints from the bounding box
The diagonal of enclosing rectangle is needed to determine the
expected span of distance dependence. The previous implementation
for SpatialPoints computed full spatial distance matrix to find
the maximum span, but this can be big job for large matrices. Now
we just find it as the diameter of the bounding box, similarly as
for non-spatial points.