save .Random.seed used in MCMC sampling
Save the internal value of .Random.seed. This allows replication
even when set.seed() was not used. We do not save the initSeed
seeds for chains, as there is no obvious way of resetting these
in the analysis. Use .RandomSeed <- hM$randSeed to reproduce the
same random sequences for the same number of chains (cannot be
reproduced across different numbers of chains).