add arg 'yshow' and fix showing data in quantitative models
Arg 'yshow' is a vector (of length 1 or longer) that gives the
values that must be included in the y-axis. So with yshow=0, zero
will always be shown, and in probit models yshow=c(0,1) shows both
zero and one.

Quantitative models data with measure="S" could contain NA and data
were not shown.