plotGradient gained main= argument similarly as other plots
main= already worked with continuous graphics that passed extras
(...) to plot. However, ggplot2 graphics were used for factor
"gradients" and main did not work there. You could add the title
by catching the plot object and adding title there (+ labs(title=main)),
but this was a bit esoteric for many users. Now we do it within
plotBeta as a Service to The People.