less obscure error message on non-ultrametric phylogenies
If tree branches hang at different heights, the tree may not be
ultrametric and cannot be inverted. Previously the error message
(from sampleMcmc) was:

Error in chol.default(Q) :
  the leading minor of order 4 is not positive definite

which was impenetrable and unhelpful. Now we at least tell that
the problem is in phylogeny data.

The check is first done in sampleMcmc so that wrong data are still
silently passsed on from model definition in Hmsc(). Would need to
add if (det(C) <= EPS) to Hmsc to catch this earlier.