centre knots in coordinates and do not give more knots than requested
Knots are found in sequence by knotDist and there can be long gap
at the upper limit, but start is always at the coordinate minimum.
Now we move the knot locations so that the distance to the  minimum
is equal to the distance to the maximum, and knots are centred
(this became obvious in vignette_4_spatial Fig. 4).

Even when the user specifies number of knots, they are found via
knotDists, but the number of intervals is one less than the number
of knots. Earlier we gave one knot more than requested for
shortest dimension.