honour start & thin in cross-validated (n-fold) predicted values
Functin computePredictedValues always returned array with original
number of samples even if user defined thin or start for a smaller
array. The real bug was that only the reduced number of samples
was calculated, but the array was filled to the original number
of samples by replicating predicted values. Discussed in issue #86
in github.