wrap every updater in try() and update params only if no try-error
Some updaters can fail, sometimes after days of sampling, and then
all results are lost. Now we wrap updaters in try() and if updater
fails, we use the old parameters.

This is the first step of handling errors. We are now loud and
error messages are printed although they are not fatal, and we do
not say where they happen. We should rather collect counts of failed
update steps and return this information after sampling. A small number
of failed updaters can be tolerated, but results are not to be trusted
if many updates fail.