collect count on updater failures and report after sampleMcmc
In parallel processing, try() errors are silenced and the user
does not learn about them. Now we collect failure counts for
every try of update. sampleChain returns only the postList and
we add failure counts there, but *remove* after reporting as these
can bring havoc to handling and pooling postList. The removal needs
a second thought, and it makes sense to retain the info on
failure counts.