knots cannot duplicate data points in GPP models: warn and list
detect and warn: will (probably) stop soon in data preparation, but
we only warn. It seems that these can sometimes survive some duplicates,
so we do not stop. However, this may be for causes where zero
distance is numerically something like 2.2e-16 (for which we don't
warn –– perhaps we should?)