Full spatial model fails when new Gradient locations duplicate data
We don't allow duplicated spatial locations (or time points), but
in predict the constructed Gradient may contain such. The sd of
point is related to distance, and with zero-distance of duplicate,
the chol(W) can be zero or tiny negative (say, -2.2e-6), and this
fails fatally. This fix assumes that all chol() failures are zero
sd cases, and we just drop adding the rnorm() term in prediction.

The problem was found with a problem data prepared and supplied
by Beb Weigel (Uni Helsinki). The same problem will occur also
in other spatical models, but only the Full model is fixed here.