move code within predN loop to a separate function
The link loops over a large number of lines and now all these lines
are made to an internal function. The goal is to parallelize this
loop. Currently there are not many other changes than this packaging,
and the function won't work. I should have changed more stages of
development, but the only really important change is that the loop
only passes one pN item of post (used as sam in function) and
predPostEta. The comments in the code tell what objects should still
passed to the function.