implement parallel processing for the prediction loop
Rough code and only works on unix-like systems (Linux, macOS).
Not documented yet, and computePredictedValues does not pass
nParallel argument – so made default nParallel=6 for testing
purposes.

This is really useful only when Yc is given and mcmcStep is set,
because this takes time. The idea is to make pcomputePredictedValues
(i.e., parallel version) to pass nParallel here when species
cross-validation was requested.