spatial weight matrices were from old data with new Gradient
Issue #135: Error: Matrices must have same dimensions in iWs + tmp1
triggered in updateEta. Looks like the problem is that the spatial
weight matrices iWs were found from the original model object, but
tmp1 and other calculations were made for the new spatial data
supplied in Gradient. If these had different number of observations,
there was an error. If these had same dimensions but spatial
coordinates changed, there was no error, but the results were wrong.

This commit inspects how to do this and only handles Full spatial
model, and even that crudely. If this is the way to go in this fix,
the code needs refactoring and clean-up.