refactor sampleMcmc to allow alternative MCMC samplers: first steps
The purpose is to separate the actual time-consuming MCMC sampling
to a dedicated function so that we can develop alternative
samplers (such as based on tensorflow). This is the first step
and won't work (except for samplePrior=TRUE): need to have a
break and save the current stage of development.