add arg engine to allow alternative or experimental MCMC samplers
This allows maintaining and developing alternative methods without
interfering with the current interface: just add a choice in
switch(engine). The choice can be non-public or public -- the only
difference is that public choices are documented in @param engine.
There is now a non-public choice "pass" that returns the
samplingObject.