lift initParList from sampler/sampleChain to sampling obj prepare
sampleMcmc has only two commands: 1) prepare sampling object for
a sampler, and 2) use a sampler. The second function was based on
otherwise non-modified internal sampleMcmc which also included
computing initial parameters for each chain. This fits better in
preparation steps so that the initialization can be used in
any alternative sampler not involving sampleChain.